requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# LLM APIs
anthropic==0.39.0
openai==1.54.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utilities
python-dateutil==2.8.2
